Fix overlay for current buster kernel 5.10.x
The overlay in the EEPROM is for kernel 4.x and soes not ork on later kernels.
The one in github (and current Raspberry Pi OS buster distribution) is for jernel 5.4.x.
It basically works , but the backlight is off -> black screen.
Use the overlay from this commit for the current 5.10.x buster kernel.

Note: TFT displays are not supported (et) in bullseye